requirements.txt
alpaca-py==0.43.2
annotated-types==0.7.0
bidict==0.23.1
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
eventlet==0.40.4
Flask==3.1.2
Flask-SocketIO==5.6.0
greenlet==3.3.0
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
msgpack==1.1.2
numpy==2.4.1
pandas==2.3.3
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-engineio==4.13.0
python-socketio==5.16.0
pytz==2025.2
requests==2.32.5
simple-websocket==1.1.0
six==1.17.0
sseclient-py==1.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
websockets==16.0
Werkzeug==3.1.5
wsproto==1.3.2
